+++ corehq/apps/dump_reload/couch/dump.py
@@ -101,7 +101,7 @@ class ToggleDumper(DataDumper):
 def _get_toggles_to_migrate(domain):
     from corehq.toggles import all_toggles, NAMESPACE_DOMAIN
     from toggle.models import Toggle
-    from toggle.shortcuts import namespaced_item
+    from corehq.toggles.shortcuts import namespaced_item
 
     domain_item = namespaced_item(domain, NAMESPACE_DOMAIN)
     usernames = set(get_all_usernames_by_domain(domain))
