--- corehq/ex-submodules/toggle/__init__.py
@@ -1,1 +1,0 @@
-from .shortcuts import toggle_enabled
