+++ corehq/motech/dhis2/tasks.py
@@ -6,7 +6,7 @@ from celery.schedules import crontab
 from celery.task import periodic_task, task
 
 from corehq.motech.utils import pformat_json
-from toggle.shortcuts import find_domains_with_toggle_enabled
+from corehq.toggles.shortcuts import find_domains_with_toggle_enabled
 
 from corehq import toggles
 from corehq.motech.dhis2.models import (
