+++ corehq/toggles/models.py
@@ -1,0 +1,1 @@
+from toggle.models import Toggle, generate_toggle_id
