+++ corehq/toggles/shortcuts.py
